requirements.txt
appdirs==1.4.3
appnope==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.3
bleach==2.0.0
certifi==2017.4.17
cffi==1.10.0
click==6.7
configparser==3.5.0
cycler==0.10.0
decorator==4.0.11
Django==1.11.1
entrypoints==0.2.2
enum34==1.1.6
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.2
functools32==3.2.3.post2
html5lib==0.999999999
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
mpmath==0.19
MySQL-python==1.2.5
nbconvert==5.1.1
nbformat==4.3.0
nose==1.3.7
notebook==5.0.0
numpy==1.12.1
packaging==16.8
pandas==0.20.1
pandocfilters==1.4.1
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pycparser==2.17
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.14.2
scandir==1.5
scipy==0.19.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.10
subprocess32==3.2.7
sympy==1.0
terminado==0.6
testpath==0.3
tornado==4.5.1
traitlets==4.3.2
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.1
widgetsnbextension==2.0.0
